--- oidc/src/components/Info.js
@@ -1,154 +1,0 @@
-import logo from './logo.svg';
-import './App.css';
-import './style.css'
-import 'bootstrap/dist/css/bootstrap.min.css';
-import React, { useEffect, useState } from "react";
-import Table from 'react-bootstrap/Table';
-import $ from "jquery";
-import { Component } from "react";
-import { BrowserRouter, Routes, Route } from 'react-router-dom';
-
-
-function Footer() {
-    return <footer class="py-4 bg-light mt-auto">
-        <div class="container-fluid px-4">
-            <div class="d-flex align-items-center justify-content-between small">
-                <div class="text-muted">Copyright &copy; Your Website 2022</div>
-               
-            </div>
-        </div>
-    </footer>
-}
-
-function Nav() {
-    return  <nav class="sb-topnav navbar navbar-expand navbar-dark bg-dark">
-            <a class="navbar-brand ps-3" href="/">MigHelper</a>
-            <button class="btn btn-link btn-sm order-1 order-lg-0 me-4 me-lg-0" id="sidebarToggle" href="#!"><i class="fas fa-bars"></i></button>
-            
-            <ul class="navbar-nav ms-auto ms-md-0 me-3 me-lg-4">
-                <li class="nav-item dropdown">
-                    
-                    <ul class="dropdown-menu dropdown-menu-end" aria-labelledby="navbarDropdown">
-                        <li><a class="dropdown-item" href="#!">Settings</a></li>
-                        <li><a class="dropdown-item" href="#!">Activity Log</a></li>
-                        <li><hr class="dropdown-divider" /></li>
-                        <li><a class="dropdown-item" href="#!">Logout</a></li>
-                    </ul>
-                </li>
-            </ul>
-        </nav>
-}
-
-function Info(){
-    return (
-        <div className="MigHelper">
-        <body class="sb-nav-fixed">
-           <Nav></Nav>
-            <div id="layoutSidenav">
-                <div id="layoutSidenav_nav">
-                    <nav class="sb-sidenav accordion sb-sidenav-dark" id="sidenavAccordion">
-                        <div class="sb-sidenav-menu">
-                            <div class="nav">
-                                <div class="sb-sidenav-menu-heading">Info</div>
-                                <a class="nav-link" href="/">
-                                    <div class="sb-nav-link-icon"><i class="fas fa-tachometer-alt"></i></div>
-                                    Infomation
-                                </a>
-                                <div class="sb-sidenav-menu-heading">Use Helper</div>
-                                <a class="nav-link collapsed" href="/mighelper" data-bs-toggle="collapse" data-bs-target="#collapseLayouts" aria-expanded="false" aria-controls="collapseLayouts">
-                                    <div class="sb-nav-link-icon"><i class="fas fa-columns"></i></div>
-                                    Migration Helper
-                                    <div class="sb-sidenav-collapse-arrow"><i class="fas fa-angle-down"></i></div>
-                                </a>
-                                <div class="collapse" id="collapseLayouts" aria-labelledby="headingOne" data-bs-parent="#sidenavAccordion">
-                                    <nav class="sb-sidenav-menu-nested nav">
-                                        <a class="nav-link" href="layout-static.html">Static Navigation</a>
-                                        <a class="nav-link" href="layout-sidenav-light.html">Light Sidenav</a>
-                                    </nav>
-                                </div>
-                                <a class="nav-link collapsed" href="/createserver" data-bs-toggle="collapse" data-bs-target="#collapsePages" aria-expanded="false" aria-controls="collapsePages">
-                                    <div class="sb-nav-link-icon"><i class="fas fa-book-open"></i></div>
-                                    Server Create Helper
-                                    <div class="sb-sidenav-collapse-arrow"><i class="fas fa-angle-down"></i></div>
-                                </a>
-                                <div class="collapse" id="collapsePages" aria-labelledby="headingTwo" data-bs-parent="#sidenavAccordion">
-                                    <nav class="sb-sidenav-menu-nested nav accordion" id="sidenavAccordionPages">
-                                        <a class="nav-link collapsed" href="#" data-bs-toggle="collapse" data-bs-target="#pagesCollapseAuth" aria-expanded="false" aria-controls="pagesCollapseAuth">
-                                            Authentication
-                                            <div class="sb-sidenav-collapse-arrow"><i class="fas fa-angle-down"></i></div>
-                                        </a>
-                                        <div class="collapse" id="pagesCollapseAuth" aria-labelledby="headingOne" data-bs-parent="#sidenavAccordionPages">
-                                            <nav class="sb-sidenav-menu-nested nav">
-                                                <a class="nav-link" href="login.html">Login</a>
-                                                <a class="nav-link" href="register.html">Register</a>
-                                                <a class="nav-link" href="password.html">Forgot Password</a>
-                                            </nav>
-                                        </div>
-                                        <a class="nav-link collapsed" href="#" data-bs-toggle="collapse" data-bs-target="#pagesCollapseError" aria-expanded="false" aria-controls="pagesCollapseError">
-                                            Error
-                                            <div class="sb-sidenav-collapse-arrow"><i class="fas fa-angle-down"></i></div>
-                                        </a>
-                                        <div class="collapse" id="pagesCollapseError" aria-labelledby="headingOne" data-bs-parent="#sidenavAccordionPages">
-                                            <nav class="sb-sidenav-menu-nested nav">
-                                                <a class="nav-link" href="401.html">401 Page</a>
-                                                <a class="nav-link" href="404.html">404 Page</a>
-                                                <a class="nav-link" href="500.html">500 Page</a>
-                                            </nav>
-                                        </div>
-                                    </nav>
-                                </div>
-                                <div class="sb-sidenav-menu-heading">Dashboard</div>
-                                <a class="nav-link" href="charts.html">
-                                    <div class="sb-nav-link-icon"><i class="fas fa-chart-area"></i></div>
-                                    Dashboard
-                                </a>
-                            </div>
-                        </div>
-                        <div class="sb-sidenav-footer">
-                            <div class="small">Logged in as:</div>
-                            Start Bootstrap
-                        </div>
-                    </nav>
-                </div>
-                <div id="layoutSidenav_content">
-                    <main>
-                        <div class="container-fluid px-4">
-                            <h1 class="mt-4">Information of MigHelper</h1>
-                            <ol class="breadcrumb mb-4">
-                                <li class="breadcrumb-item active">Information</li>
-                            </ol>
-                            <p1>안녕하세요. 제 4회 OIDC에 참가한 팀 정통입니다.</p1>
-                            <br></br>
-                            <p1>팀 정통이 해당 서비스를 제공하고자 하는 목적은 다음과 같습니다.</p1>
-                            <br></br>
-                            <br></br>
-                            <p1>많은 기업들이 기존의 On-premise 환경을 Cloud 환경으로 Lift and Shift를 진행하고 있고, 이러한 migration을 위해 많은 기업들이 MSP 기업이나 Cloud 전문가에게 거액의 돈을 주고 migration을 의뢰하여 진행하고 있습니다. 이러한 돈이 부담되는 migration 의뢰를 MSP 기업의 도움 없이, 기업이 직접 웹서비스를 활용하여 기업이 보유한 자원에 따른 Cloud 서버 및 규모와 비용 산정에 도움 주고자 함이 목적입니다.</p1>
-                            <br></br>
-                            <p1>또한 Cloud를 이용하는 기업의 경우, CSP에서 제공하는 Console을 활용해 Cloud 내에서 진행되는 워크로드들을 일일히 확인해야 하는데, 현재 Cloud 내에서 진행되는 워크로드를 한 눈으로 모니터링 할 수 있는 대시보드를 제공하여 일일히 모든 서비스를 확인해야 하는 번거로움을 줄이고, 보안 취약점을 제시하여 기업의 Cloud 환경의 보안을 강화 하고자 함이 목적입니다.</p1>
-                            <br></br>
-                            <br></br>
-                            <p1>우리가 제공하는 서비스는 아래와 같습니다.</p1>
-                            <br></br>
-                            <br></br>
-                            <ul>
-                                <li>MigHelper : 클라우드로의 마이그레이션 시 솔루션 제공</li>
-                                <li>Create Server Helper : 원하는 사양에 맞는 클라우드 추천</li>
-                                <li>Dashboard : 로그인 후 이용가능한 기능으로 자신의 클라우드 상태를 확인할 수 있는 대시보드 제공</li>
-                                <li>Plus 기능 : 보안 거버넌스 제공</li>
-                            </ul>
-                        </div>
-                    </main>
-                 <Footer></Footer>
-                </div>
-            </div>
-            <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js" crossorigin="anonymous"></script>
-            <script src="js/scripts.js"></script>
-            <script src="https://cdnjs.cloudflare.com/ajax/libs/Chart.js/2.8.0/Chart.min.js" crossorigin="anonymous"></script>
-            <script src="assets/demo/chart-area-demo.js"></script>
-            <script src="assets/demo/chart-bar-demo.js"></script>
-            <script src="https://cdn.jsdelivr.net/npm/simple-datatables@latest" crossorigin="anonymous"></script>
-            <script src="js/datatables-simple-demo.js"></script>
-        </body>
-        </div>
-      );
-}export default Info;
