+++ oidc/src/components/footer.js
@@ -1,0 +1,19 @@
+import React from 'react';
+//import Layout from './component/layout';
+import './App.css';
+import './style.css'
+
+function Footer() {
+    return (
+        <footer class="py-4 bg-light mt-auto">
+            <div class="container-fluid px-4">
+                <div class="d-flex align-items-center justify-content-between small">
+                    <div class="text-muted">Copyright &copy; Your Website 2022</div>
+
+                </div>
+            </div>
+        </footer>
+    );
+}
+
+export default Footer;
